requirements.txt
torch>=2.0.0
numpy>=1.19.2
pandas>=1.1.3
matplotlib>=3.3.2
seaborn>=0.11.0
biopython>=1.78
streamlit>=1.10.0
plotly>=4.14.3
scikit-learn>=0.24.0
tqdm>=4.50.2
scipy>=1.7.0
